+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [0.0.1] - 2023-08-03
+## [0.0.1] - 2024-06-20
 
 ### Added
 
+ - dataTransfer module, POC for REST pull artifact
+ - duplicate TransferProcess and ContractNegotiation with new status
+ - DataService, update method
+ 
 ### Changed
 
 ### Removed